requirements.txt
# Core dependencies for Resume Autofill System
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Resume parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# LLM integration
openai==1.3.7
anthropic==0.7.8
pydantic==2.5.0

# Data processing
# pandas==2.1.4  # Not compatible with Python 3.13
# numpy==1.25.2  # Not compatible with Python 3.13

# Web scraping and automation
lxml==4.9.3
webdriver-manager==4.0.1

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Web framework for extension backend
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
